@@ -1,6 +1,6 @@
 {
   "name": "portfolio",
-  "version": "1.1.0",
+  "version": "1.2.0",
   "description": "Software Engineering | Developer Portfolio",
   "private": true,
   "author": {
@@ -11,6 +11,7 @@
     "portfolio",
     "full-stack",
     "svelte",
+    "svelte5",
     "sveltekit",
     "typescript",
     "tailwindcss",
@@ -20,50 +21,50 @@
     "type": "git",
     "url": "https://github.com/rawat9/portfolio"
   },
-  "homepage": "https://anuragrawat.tech/",
+  "homepage": "https://anuragsrawat.vercel.app/",
   "scripts": {
     "dev": "vite dev",
     "build": "vite build",
     "preview": "vite preview",
     "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
     "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
-    "lint": "prettier --plugin-search-dir . --check . && eslint .",
-    "format": "prettier --plugin-search-dir . --write ."
+    "lint": "prettier . --check . && eslint .",
+    "format": "prettier . --write ."
   },
   "devDependencies": {
-    "@iconify/svelte": "4.0.1",
-    "@sveltejs/adapter-vercel": "5.3.0",
-    "@sveltejs/kit": "2.5.6",
-    "@sveltejs/vite-plugin-svelte": "^3.0.1",
+    "@iconify/svelte": "4.0.2",
+    "@sveltejs/adapter-vercel": "5.4.6",
+    "@sveltejs/kit": "2.7.3",
+    "@sveltejs/vite-plugin-svelte": "4.0.0",
     "@types/nodemailer": "6.4.14",
+    "@types/reveal.js": "5.0.4",
     "@typescript-eslint/eslint-plugin": "7.7.0",
     "@typescript-eslint/parser": "7.7.0",
     "autoprefixer": "10.4.16",
     "eslint": "8.56.0",
     "eslint-config-prettier": "9.1.0",
-    "eslint-plugin-svelte": "2.35.1",
+    "eslint-plugin-svelte": "2.46.0",
     "postcss": "8.4.32",
-    "prettier": "3.2.5",
-    "prettier-plugin-svelte": "3.2.3",
-    "svelte": "4.2.14",
-    "svelte-check": "3.6.9",
+    "prettier": "3.3.3",
+    "prettier-plugin-svelte": "3.2.7",
+    "svelte": "5.1.3",
+    "svelte-check": "4.0.5",
     "svelte-typewriter": "3.2.3",
-    "tailwindcss": "3.4.3",
-    "tslib": "2.6.2",
-    "typescript": "5.4.5",
-    "vite": "5.2.11"
+    "tailwindcss": "3.4.14",
+    "tslib": "2.8.0",
+    "typescript": "5.6.3",
+    "vite": "5.4.10"
   },
   "type": "module",
   "dependencies": {
-    "@fontsource/geist-sans": "5.0.3",
-    "@types/reveal.js": "5.0.3",
-    "bits-ui": "0.13.0",
+    "@fontsource/geist-sans": "5.1.0",
+    "bits-ui": "1.0.0-next.29",
     "clsx": "2.1.0",
     "nodemailer": "6.9.9",
     "reveal.js": "5.1.0",
     "svelte-email": "0.0.4",
-    "svelte-french-toast": "1.2.0",
-    "tailwind-merge": "2.2.0",
-    "tailwind-variants": "0.1.19"
+    "svelte-sonner": "^0.3.28",
+    "tailwind-merge": "2.5.4",
+    "tailwind-variants": "0.2.1"
   }
 }
