@@ -1,21 +1,19 @@
 <script lang="ts">
   import { Tabs as TabsPrimitive } from 'bits-ui';
-  import { cn } from '$lib/utils';
+  import { cn } from '$lib/utils.js';
 
-  type $$Props = TabsPrimitive.ContentProps;
-
-  let className: $$Props['class'] = undefined;
-  export let value: $$Props['value'];
-  export { className as class };
+  let {
+    ref = $bindable(null),
+    class: className,
+    ...restProps
+  }: TabsPrimitive.ContentProps = $props();
 </script>
 
 <TabsPrimitive.Content
+  bind:ref
   class={cn(
-    'mt-2 ring-offset-background focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2',
+    'ring-offset-background focus-visible:ring-ring mt-2 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-offset-2',
     className
   )}
-  {value}
-  {...$$restProps}
->
-  <slot />
-</TabsPrimitive.Content>
+  {...restProps}
+/>
