@@ -1,28 +1,21 @@
 <script lang="ts">
   import { Tooltip as TooltipPrimitive } from 'bits-ui';
-  import { cn, flyAndScale } from '$lib/utils';
+  import { cn } from '$lib/utils.js';
 
-  type $$Props = TooltipPrimitive.ContentProps;
-
-  let className: $$Props['class'] = undefined;
-  export let sideOffset: $$Props['sideOffset'] = 4;
-  export let transition: $$Props['transition'] = flyAndScale;
-  export let transitionConfig: $$Props['transitionConfig'] = {
-    y: 10,
-    duration: 100
-  };
-  export { className as class };
+  let {
+    ref = $bindable(null),
+    class: className,
+    sideOffset = 4,
+    ...restProps
+  }: TooltipPrimitive.ContentProps = $props();
 </script>
 
 <TooltipPrimitive.Content
-  {transition}
-  {transitionConfig}
+  bind:ref
   {sideOffset}
   class={cn(
-    'z-50 overflow-hidden rounded-md border bg-popover px-3 py-1.5 text-sm text-popover-foreground shadow-md',
+    'bg-popover text-popover-foreground z-50 overflow-hidden rounded-md border px-3 py-1.5 text-sm shadow-md',
     className
   )}
-  {...$$restProps}
->
-  <slot />
-</TooltipPrimitive.Content>
+  {...restProps}
+/>
