@@ -37,7 +37,7 @@
       placeholder="Your message"
       required
       maxLength={500}
-    />
+    ></textarea>
     <button
       type="submit"
       class="group flex items-center justify-center gap-2 h-[3rem] w-[8rem] bg-gray-900 text-white rounded-full outline-none transition-all hover:bg-gray-950 active:scale-105 dark:bg-white dark:bg-opacity-10 disabled:scale-100 disabled:bg-opacity-65"
